Manage DJI + Raceband together
During races, we do not only have R-band pilots.
Most of the time, few DJI pilots are also there.

So far, managing such pilots was painful as it needed a manual change
either in nodes settings or heat pilot seats in order to have correct
DJI pilots configured on correct nodes.

With this new plugin and associated patches, DJI pilots are
automatically assigned to a D band node if available.

This is mostly useful on first heat, when no history has been recorded
yet. On subsequent heats, the current algorithm would reuse the previous
frequencies.

This commit is also introducing a new configuration for 8 nodes RH with:
- R1378 + D1367

D1367 are the equivalent of R1378 for DJI pilots.
R6 has been discarded on purpose because it's DJI public channel (used
by DJI VRX).

The current patch will discard the frequencies from R/D if any D/R
related freq is already affected to a seat.
E.G. if a pilot is affected R1, then D1 will be discarded from available
seats so that it never get affected to a DJI pilot.

Fix bug GH-468

Signed-off-by: Arnaud M <arnaud.choupinette@mailops.fr>