Fix: use class id as param to acutally commit to DB
The ranking was not commited correctly into database somehow due to
DB context not beeing the same when getting the object and updating it.

When using the class id as param to both clear and set ranking
functions, it forces the functions to get the raceClass using the same
db session as the one used for commit.

Signed-off-by: Arnaud M <arnaud.choupinette@mailops.fr>